backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0

openai>=1.40.0
httpx>=0.27.0

faiss-cpu==1.8.0.post1
numpy>=1.26.0

gitpython==3.1.43
python-multipart==0.0.9
sentence-transformers==3.0.1